requirements.txt
eth-brownie==1.16.0
web3==5.22.0
python-dotenv==0.16.0
pytest==6.2.4